+++ backend/DatabaseManager.h
@@ -3,27 +3,22 @@
 #include <QString>
 #include <QVariantList>
 
-// SQLite 기반 이벤트 로그 매니저 (싱글톤)
-// QSqlDatabase 기본 커넥션을 사용하므로 단일 스레드에서만 호출할 것
+// SQLite event-log manager (singleton, single-threaded)
 class DatabaseManager
 {
 public:
     static DatabaseManager& instance();
 
-    // 앱 시작 시 1회 호출 — DB 파일 생성 및 테이블 초기화
+    // Call once at startup — opens DB and creates tables
     bool init(const QString& path = "facility.db");
 
-    // ── 장비 목록 영구 저장 ────────────────────────────────────────────────
-    // 저장된 장비 목록을 order_index 순으로 반환
+    // Equipment persistence
     QVariantList loadEquipment() const;
-    // 장비 추가 (order_index = 기존 최대 + 1)
     void saveNewEquipment(const QString& id, const QString& name, const QString& imageSource);
-    // 장비 삭제
     void deleteEquipment(const QString& id);
-    // 장비 정보 수정
     void updateEquipment(const QString& id, const QString& name, const QString& imageSource);
 
-    // ── 상태 이벤트 로그 ───────────────────────────────────────────────────
+    // State event log
     QVariantList queryStateEvents(const QString& equipmentId, int limit = 50) const;
     void clearEquipmentEvents(const QString& equipmentId);
     void insertStateEvent(
@@ -31,9 +26,9 @@ public:
         const QString& equipmentName,
         const QString& state,
         const QString& controlStatus,
-        float temperature,
-        float power,
-        qint64 timestampMs   // original log timestamp
+        int temperature,
+        int power,
+        qint64 timestampMs
     );
 
 private:
